requirements.txt
django
djangorestframework
parameterized==0.7.0
PyDispatcher==2.0.5
